Field polishing and improvements

+++ Joinrpg/Controllers/GameFieldController.cs
@@ -20,90 +20,91 @@ namespace JoinRpg.Web.Controllers
 
     private ActionResult ReturnToIndex(Project project)
     {
-      return RedirectToAction("Index", new {project.ProjectId});
+      return RedirectToAction("Index", new { project.ProjectId });
     }
 
+
     [HttpGet]
-    [Authorize]
     // GET: GameFields
     public async Task<ActionResult> Index(int projectId)
     {
       var project1 = await ProjectRepository.GetProjectAsync(projectId);
-      return AsMaster(project1, pa => pa.CanChangeFields) ?? View(project1);
+      return AsMaster(project1, pa => pa.CanChangeFields) ?? View(new GameFieldListViewModel()
+      {
+        ProjectId = project1.ProjectId,
+        Items = project1.AllProjectFields.Select(pf => new GameFieldEditViewModel(pf))
+      });
     }
 
+    [HttpGet]
     // GET: GameFields/Create
     public async Task<ActionResult> Create(int projectId)
     {
       var project1 = await ProjectRepository.GetProjectAsync(projectId);
-      return AsMaster(project1, pa => pa.CanChangeFields) ?? View(new ProjectCharacterField() {ProjectId = projectId});
+      return AsMaster(project1, pa => pa.CanChangeFields) ?? View(new GameFieldCreateViewModel() {ProjectId = projectId});
     }
 
     // POST: GameFields/Create
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public ActionResult Create(ProjectCharacterField field)
+    public async Task<ActionResult> Create(GameFieldCreateViewModel viewModel)
     {
-      var project1 = ProjectRepository.GetProject(field.ProjectId);
-      return AsMaster(project1, pa => pa.CanChangeFields) ?? ((Func<Project, ActionResult>) (project =>
+      var project = await ProjectRepository.GetProjectAsync(viewModel.ProjectId);
+      var error = AsMaster(project, pa => pa.CanChangeFields);
+      if (error != null)
       {
-        try
-        {
-          field.IsActive = true;
-          ProjectService.AddCharacterField(field);
+        return error;
+      }
+      try
+      {
+        await ProjectService.AddCharacterField(project.ProjectId, CurrentUserId, viewModel.FieldType, viewModel.Name,
+          viewModel.FieldHint,
+          viewModel.CanPlayerEdit, viewModel.CanPlayerView, viewModel.IsPublic);
 
-          return ReturnToIndex(project);
-        }
-        catch
-        {
-          return View();
-        }
-      }))(project1);
+        return ReturnToIndex(project);
+      }
+      catch
+      {
+        return View();
+      }
     }
 
+    [HttpGet]
     // GET: GameFields/Edit/5
     public ActionResult Edit(int projectId, int projectCharacterFieldId)
     {
-      return WithGameFieldAsMaster(projectId, projectCharacterFieldId, (project, field) =>
-      {
-        var viewModel = new GameFieldEditViewModel()
-        {
-          CanPlayerView = field.CanPlayerView,
-          CanPlayerEdit = field.CanPlayerEdit,
-          FieldHint = field.FieldHint,
-          FieldId = field.ProjectCharacterFieldId,
-          IsPublic = field.IsPublic,
-          Name = field.FieldName,
-          ProjectId = field.ProjectId
-        };
-        return View(viewModel);
-      });
+      return WithGameFieldAsMaster(projectId, projectCharacterFieldId,
+        (project, field) => View(new GameFieldEditViewModel(field)));
     }
 
     // POST: GameFields/Edit/5
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public ActionResult Edit(GameFieldEditViewModel viewModel)
+    public async Task<ActionResult> Edit(GameFieldEditViewModel viewModel)
     {
-      return WithGameFieldAsMaster(viewModel.ProjectId, viewModel.FieldId, (project, field) =>
-      {
-        try
-        {
-          ProjectService.UpdateCharacterField(project.ProjectId, field.ProjectCharacterFieldId, viewModel.Name,
-            viewModel.FieldHint,
-            viewModel.CanPlayerEdit, viewModel.CanPlayerView, viewModel.IsPublic);
+      var project = await ProjectRepository.GetProjectAsync(viewModel.ProjectId);
+      var field = project.AllProjectFields.SingleOrDefault(e => e.ProjectCharacterFieldId == viewModel.ProjectCharacterFieldId);
 
-          return ReturnToIndex(project);
-        }
-        catch
-        {
-          return View(viewModel);
-        }
-      });
+      var error = AsMaster(field, pa => pa.CanChangeFields);
+      if (error != null)
+      {
+        return error;
+      }
+      try
+      {
+        await ProjectService.UpdateCharacterField(project.ProjectId, field.ProjectCharacterFieldId, viewModel.Name,
+          viewModel.FieldHint,
+          viewModel.CanPlayerEdit, viewModel.CanPlayerView, viewModel.IsPublic);
+
+        return ReturnToIndex(project);
+      }
+      catch
+      {
+        return View(viewModel);
+      }
     }
 
     [HttpGet]
-    [Authorize]
     // GET: GameFields/Delete/5
     public ActionResult Delete(int projectId, int projectCharacterFieldId)
     {
@@ -114,13 +115,20 @@ namespace JoinRpg.Web.Controllers
     // POST: GameFields/Delete/5
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public ActionResult Delete(int projectId, int projectCharacterFieldId, FormCollection collection)
+    public async Task<ActionResult> Delete(int projectId, int projectCharacterFieldId, FormCollection collection)
     {
-      return WithGameFieldAsMaster(projectId, projectCharacterFieldId, (project, field) =>
+      var project = await ProjectRepository.GetProjectAsync(projectId);
+      var field = project.AllProjectFields.SingleOrDefault(e => e.ProjectCharacterFieldId == projectCharacterFieldId);
+
+      var error = AsMaster(field, pa => pa.CanChangeFields);
+      if (error != null)
+      {
+        return error;
+      }
       {
         try
         {
-          ProjectService.DeleteField(field.ProjectCharacterFieldId);
+          await ProjectService.DeleteField(field.ProjectCharacterFieldId);
 
           return ReturnToIndex(project);
         }
@@ -128,7 +136,7 @@ namespace JoinRpg.Web.Controllers
         {
           return View(field);
         }
-      });
+      }
 
 
     }
